+++ breeze-core/src/main/kotlin/com/windea/breezeframework/core/annotations/api/ExplicitApi.kt
@@ -2,4 +2,5 @@ package com.windea.breezeframework.core.annotations.api
 
 /**需要显式使用的api。*/
 @MustBeDocumented
+@Retention(AnnotationRetention.BINARY)
 internal annotation class ExplicitApi