+++ breeze-core/src/main/kotlin/com/windea/breezeframework/core/annotations/api/InfiniteLoopApi.kt
@@ -2,4 +2,5 @@ package com.windea.breezeframework.core.annotations.api
 
 /**可能引起无限循环的api。*/
 @MustBeDocumented
+@Retention(AnnotationRetention.BINARY)
 annotation class InfiniteLoopApi