+++ breeze-core/src/main/kotlin/com/windea/breezeframework/core/annotations/api/LowPerformanceApi.kt
@@ -2,4 +2,5 @@ package com.windea.breezeframework.core.annotations.api
 
 /**可能执行低性能操作的api。*/
 @MustBeDocumented
+@Retention(AnnotationRetention.BINARY)
 annotation class LowPerformanceApi