+++ breeze-core/src/main/kotlin/com/windea/breezeframework/core/annotations/api/StableApi.kt
@@ -2,4 +2,5 @@ package com.windea.breezeframework.core.annotations.api
 
 /**稳定的api。*/
 @MustBeDocumented
+@Retention(AnnotationRetention.BINARY)
 annotation class StableApi