+++ breeze-core/src/main/kotlin/com/windea/breezeframework/core/annotations/marks/NotImplemented.kt
@@ -3,8 +3,9 @@ package com.windea.breezeframework.core.annotations.marks
 import kotlin.annotation.AnnotationTarget.*
 
 /**标注为未实现的项。*/
-@Target(CLASS, PROPERTY, FUNCTION, FILE)
 @MustBeDocumented
+@Retention(AnnotationRetention.BINARY)
+@Target(CLASS, PROPERTY, FUNCTION, FILE)
 annotation class NotImplemented(
-	val message: String = "Not implemented."
+	val value: String = "Not implemented."
 )