+++ breeze-core/src/main/kotlin/com/windea/breezeframework/core/annotations/marks/NotIncluded.kt
@@ -3,8 +3,9 @@ package com.windea.breezeframework.core.annotations.marks
 import kotlin.annotation.AnnotationTarget.*
 
 /**标注为未包含的项。*/
-@Target(CLASS, PROPERTY, FUNCTION, FILE)
 @MustBeDocumented
+@Retention(AnnotationRetention.BINARY)
+@Target(CLASS, PROPERTY, FUNCTION, FILE)
 annotation class NotIncluded(
-	val message: String = "Not included."
+	val value: String = "Not included."
 )