+++ breeze-core/src/main/kotlin/com/windea/breezeframework/core/annotations/marks/NotRecommended.kt
@@ -3,8 +3,9 @@ package com.windea.breezeframework.core.annotations.marks
 import kotlin.annotation.AnnotationTarget.*
 
 /**标注为不推荐的项。*/
-@Target(CLASS, PROPERTY, FUNCTION, FILE)
 @MustBeDocumented
+@Retention(AnnotationRetention.BINARY)
+@Target(CLASS, PROPERTY, FUNCTION, FILE)
 annotation class NotRecommended(
-	val message: String = "Not recommended."
+	val value: String = "Not recommended."
 )