+++ breeze-core/src/main/kotlin/com/windea/breezeframework/core/annotations/marks/NotSuitable.kt
@@ -3,8 +3,9 @@ package com.windea.breezeframework.core.annotations.marks
 import kotlin.annotation.AnnotationTarget.*
 
 /**标注为不适用的项。*/
-@Target(CLASS, PROPERTY, FUNCTION, FILE)
 @MustBeDocumented
+@Retention(AnnotationRetention.BINARY)
+@Target(CLASS, PROPERTY, FUNCTION, FILE)
 annotation class NotSuitable(
-	val message: String = "Not suitable."
+	val value: String = "Not suitable."
 )