+++ breeze-core/src/main/kotlin/com/windea/breezeframework/core/annotations/marks/NotUsable.kt
@@ -3,8 +3,9 @@ package com.windea.breezeframework.core.annotations.marks
 import kotlin.annotation.AnnotationTarget.*
 
 /**标注为不可用的项。*/
-@Target(CLASS, PROPERTY, FUNCTION, FILE)
 @MustBeDocumented
+@Retention(AnnotationRetention.BINARY)
+@Target(CLASS, PROPERTY, FUNCTION, FILE)
 annotation class NotUsable(
-	val message: String = "Not usable."
+	val value: String = "Not usable."
 )