+++ breeze-core/src/main/kotlin/com/windea/breezeframework/core/annotations/messages/Description.kt
@@ -7,7 +7,6 @@ import org.jetbrains.annotations.*
 /**本地化描述。*/
 @MustBeDocumented
 @Repeatable
-@Retention(AnnotationRetention.RUNTIME)
 annotation class Description(
 	@Language("Markdown")
 	@Nls