+++ breeze-core/src/main/kotlin/com/windea/breezeframework/core/annotations/messages/Name.kt
@@ -6,7 +6,6 @@ import org.jetbrains.annotations.*
 /**本地化名字。*/
 @MustBeDocumented
 @Repeatable
-@Retention(AnnotationRetention.RUNTIME)
 annotation class Name(
 	@Nls
 	val text: String,