+++ breeze-core/src/main/kotlin/com/windea/breezeframework/core/annotations/messages/Tags.kt
@@ -6,7 +6,6 @@ import org.jetbrains.annotations.*
 /**本地化标签。*/
 @MustBeDocumented
 @Repeatable
-@Retention(AnnotationRetention.RUNTIME)
 annotation class Tags(
 	@Nls
 	val text: Array<String>,